refactor[properties]: refatorando alguns diretorios
