+++ documentos-pessoais-digitalizados/build.gradle.kts
@@ -23,7 +23,7 @@ dependencies {
 	implementation("io.minio:minio:8.5.17")
 	testImplementation("org.springframework.boot:spring-boot-starter-test")
 	testRuntimeOnly("org.junit.platform:junit-platform-launcher")
-	runtimeOnly("org.postgresql:postgresql")
+	runtimeOnly("com.h2database:h2")
 }
 
 tasks.withType<Test> {
