+++ documentos-pessoais-digitalizados/src/main/java/br/org/apae/documentos_pessoais_digitalizados/api/dto/req/PersonalDocumentReqDTO.java
@@ -1,4 +1,4 @@
-package br.org.apae.documentos_pessoais_digitalizados.api.dtos.req;
+package br.org.apae.documentos_pessoais_digitalizados.api.dto.req;
 
 import java.util.List;
 import java.util.UUID;
