+++ documentos-pessoais-digitalizados/src/main/java/br/org/apae/documentos_pessoais_digitalizados/api/dto/res/PersonalDocumentURL.java
@@ -1,0 +1,4 @@
+package br.org.apae.documentos_pessoais_digitalizados.api.dto.res;
+
+public record PersonalDocumentURL(String url) {
+}
