--- documentos-pessoais-digitalizados/src/main/java/br/org/apae/documentos_pessoais_digitalizados/api/dtos/res/PersonalDocumentURL.java
@@ -1,4 +1,0 @@
-package br.org.apae.documentos_pessoais_digitalizados.api.dtos.res;
-
-public record PersonalDocumentURL(String url) {
-}
