+++ documentos-pessoais-digitalizados/src/main/java/br/org/apae/documentos_pessoais_digitalizados/domain/model/PersonalDocumentType.java
@@ -1,4 +1,4 @@
-package br.org.apae.documentos_pessoais_digitalizados.domain.models;
+package br.org.apae.documentos_pessoais_digitalizados.domain.model;
 
 public enum PersonalDocumentType {
     CPF("cpf"),
