+++ documentos-pessoais-digitalizados/src/main/resources/application.properties
@@ -1,1 +1,6 @@
 spring.application.name=documentos-pessoais-digitalizados
+
+storage.service.base.url=""
+minio.endpoint=""
+minio.access-key=""
+minio.secret-key=""
